Agent Wallets - Move build_ttl_policy to policies/ttl.py
